+++ src/createLongArray.ts
